+++ test/libsolidity/syntaxTests/string/concat/string_concat_named_arguments.sol
@@ -1,0 +1,7 @@
+contract C {
+    function f() public pure {
+        string.concat({x: "abc"});
+    }
+}
+// ----
+// TypeError 7173: (52-77): Named arguments cannot be used for the string.concat function call.